requirements.txt
streamlit==1.32.0
numpy
pandas
gtts
google-generativeai
Pillow
flatlib
lunar_python
